wip: implement host facing on test chip
