+++ src/ariel-os-log-transport-driver/Cargo.toml
@@ -7,11 +7,19 @@ repository.workspace = true
 license.workspace = true
 
 [dependencies]
+ariel-os-boards = { workspace = true, optional = true }
+ariel-os-embassy-common = { workspace = true, optional = true }
 ariel-os-hal = { workspace = true }
 ariel-os-log = { workspace = true, features = ["custom-transport"] }
 embassy-executor = { workspace = true }
+embassy-futures = { workspace = true, optional = true }
+embassy-sync = { workspace = true, optional = true }
+embedded-io-async-06 = { package = "embedded-io-async", version = "0.6", optional = true }
+static_cell = { workspace = true, optional = true }
 
 [features]
+
+uart = ["dep:ariel-os-boards", "dep:ariel-os-embassy-common", "dep:embassy-futures", "dep:embassy-sync","dep:embedded-io-async-06", "dep:static_cell"]
 
 _test = []
