+++ src/ariel-os-log-transport-driver/src/lib.rs
@@ -5,6 +5,10 @@
 #![deny(missing_docs)]
 
 cfg_select! {
+    feature = "uart" => {
+        mod uart;
+        pub use uart::init;
+    }
     _ => {
         mod dummy;
         pub use dummy::init;
