+++ src/ariel-os-log-transport-driver/src/uart.rs
@@ -1,11 +1,14 @@
 use embassy_executor::Spawner;
+use embassy_sync::{blocking_mutex::raw::CriticalSectionRawMutex, mutex::Mutex};
+use embedded_io_async_06::Write;
 use static_cell::ConstStaticCell;
 
 use ariel_os_embassy_common::uart::{Assignment, Baudrate};
 use ariel_os_hal::hal::{OptionalPeripherals, TakePeripherals, uart::Uart};
 
 type UartAssignment = ariel_os_boards::pins::HOST_FACING_UART;
-pub type WriterType = Uart<'static>;
+
+static WRITER: Mutex<CriticalSectionRawMutex, Option<Uart<'static>>> = Mutex::new(None);
 
 static RX_BUF: ConstStaticCell<[u8; 32]> = ConstStaticCell::new([0u8; 32]);
 static TX_BUF: ConstStaticCell<[u8; 32]> = ConstStaticCell::new([0u8; 32]);
@@ -24,5 +27,24 @@ pub fn init(mut peripherals: &mut OptionalPeripherals, _spawner: Spawner) {
 
     let uart = <UartAssignment as Assignment>::Device::new(rx, tx, rx_buf, tx_buf, config).unwrap();
 
-    super::init_writer(uart);
+    let mut writer = WRITER.try_lock().unwrap();
+    writer.replace(uart);
+
+    ariel_os_log::custom_transport::register_transport_functions(write_bytes, flush);
+}
+
+pub(crate) fn write_bytes(bytes: &[u8]) {
+    if let Ok(mut opt) = WRITER.try_lock()
+        && let Some(writer) = opt.as_mut()
+    {
+        let _ = embassy_futures::block_on(writer.write_all(bytes));
+    }
+}
+
+pub(crate) fn flush() {
+    if let Ok(mut opt) = WRITER.try_lock()
+        && let Some(writer) = opt.as_mut()
+    {
+        let _ = embassy_futures::block_on(writer.flush());
+    }
 }
