--- src/ariel-os-log-transport/Cargo.toml
@@ -1,41 +1,0 @@
-[package]
-name = "ariel-os-log-transport"
-version = "0.5.0"
-edition.workspace = true
-rust-version.workspace = true
-repository.workspace = true
-license.workspace = true
-
-[dependencies]
-ariel-os-debug = { workspace = true, optional = true }
-ariel-os-boards = { workspace = true, optional = true }
-ariel-os-embassy-common = { workspace = true }
-ariel-os-hal = { workspace = true }
-ariel-os-log = { workspace = true }
-ariel-os-utils = { workspace = true, optional = true }
-critical-section = { workspace = true, optional = true }
-embassy-executor = { workspace = true }
-embassy-futures = { workspace = true, optional = true }
-embassy-sync = { workspace = true, optional = true }
-embedded-io-async = { workspace = true, optional = true }
-static_cell = { workspace = true, optional = true }
-
-[features]
-
-external = ["ariel-os-log/custom-transport"]
-
-writer = ["dep:embedded-io-async", "dep:embassy-futures", "dep:embassy-sync"]
-
-logging-over-uart = [
-  "dep:ariel-os-boards",
-  "dep:static_cell",
-  "external",
-  "writer",
-  "ariel-os-hal/uart",
-]
-logging-over-usb = ["external"]
-
-_test = []
-
-[lints]
-workspace = true
