+++ src/ariel-os/Cargo.toml
@@ -27,6 +27,7 @@ ariel-os-embassy = { path = "../ariel-os-embassy" }
 ariel-os-hal = { workspace = true }
 ariel-os-identity = { workspace = true }
 ariel-os-log = { workspace = true }
+ariel-os-log-transport-driver = { workspace = true, optional = true }
 ariel-os-macros = { path = "../ariel-os-macros" }
 ariel-os-nrf = { path = "../ariel-os-nrf", optional = true }
 ariel-os-power = { path = "../ariel-os-power" }
@@ -221,7 +222,7 @@ no-boards = ["ariel-os-boards/no-boards", "ariel-os-embassy/no-boards"]
 # These are selected by laze
 # NOTE: `time` is only needed on RP.
 logging-over-uart = [
-  "ariel-os-transport-driver/uart",
+  "ariel-os-log-transport-driver/uart",
   "log-transport-driver",
   "time",
   "uart",
